Changed the wrong name 'SignUpScreen' to SignInScreen' in my js file. Added social media logos. Finished SignInScreen.js. Could be changed in the future.
